test: separate failed buffer allocation uint32array check
